Remove privacyidea/lib/policymatch.py
